Add framework to run quickstarter tests
This replaces the test setup in ods-quickstarters.

Instead of having the test logic in the quickstarter repository to test, we
define the tests once in ods-core, and all quickstarters can be tested
using that framework by just bringing a "testdata" directory which
contains the steps to execute, and the expected golden files.

This has several advantages:
* It makes it easier to test quickstarters not located in
ods-quickstarters
* It avoids to duplicate code (e.g. shell scripts) between ods-core and
ods-quickstarters
* It makes the dependency obvious - previously one needed to run the
tests in ods-core first in order to run the ones in ods-quickstarters.
This is now more obvious and can be enforced by tooling.

Apart from this, the PR addresses all complaints by the linter (and
runs linting during CI now), and generally improves the Go code as well as the
bash scripts. Note though that the added bash scripts are just copied
from ods-quickstarters and not touched for the most part to keep impact
low.

+++ tests/verify.sh
@@ -11,6 +11,11 @@ OPENSHIFT_APPS_BASEDOMAIN=$(${ODS_CORE_DIR}/scripts/get-config-param.sh OPENSHIF
 export PROVISION_API_HOST=https://prov-app-${ODS_NAMESPACE}${OPENSHIFT_APPS_BASEDOMAIN}
 echo "PROVISION_API_HOST = ${PROVISION_API_HOST}"
 
+if ! oc whoami &> /dev/null; then
+    echo "You need to login to OpenShift to run the tests"
+    exit 1
+fi
+
 if [ -f test-results.txt ]; then
     rm test-results.txt
 fi
